refactor(bridge-core): 리팩토링 및 경로 설정 추가
- tsconfig.json에 baseUrl 및 paths 추가하여 모듈 경로 간소화
- 여러 도메인 및 기능에 대한 타입 및 RPC 스키마 추가
- 기존의 error, event-types, methods, rpc-schema, types 파일 삭제 및 새로운 구조로 통합
- 새로운 도메인 구조에 따라 각 도메인에 대한 타입 및 RPC 스키마 정의
- BridgeError 및 관련 코드 리팩토링으로 에러 처리 개선
- utils에서 ID 생성 및 JSON 파싱 유틸리티 함수 추가
- 각 도메인에 대한 이벤트 맵 및 타입 정의 추가
