+++ packages/bridge-core/src/constants/index.ts
@@ -1,0 +1,1 @@
+export { BRIDGE_VERSION } from './version';
