+++ packages/bridge-core/src/constants/version.ts
@@ -1,0 +1,4 @@
+// @TODO: 버전 관리는 이렇게 안했으면 함
+const BRIDGE_VERSION = '1.0.0';
+
+export { BRIDGE_VERSION };
