+++ packages/bridge-core/src/domains/app-version/index.ts
@@ -1,0 +1,2 @@
+export type { AppVersionRPCSchema } from './rpc-schema';
+export type { GetAppVersionResult } from './types';
