+++ packages/bridge-core/src/domains/app-version/rpc-schema.ts
@@ -1,0 +1,11 @@
+import { METHODS } from '@/rpc';
+import { type GetAppVersionResult } from './types';
+
+interface AppVersionRPCSchema {
+  [METHODS.getAppVersion]: {
+    params: {};
+    result: GetAppVersionResult;
+  };
+}
+
+export type { AppVersionRPCSchema };
