+++ packages/bridge-core/src/domains/app-version/types.ts
@@ -1,0 +1,5 @@
+type GetAppVersionResult = {
+  version: string;
+};
+
+export type { GetAppVersionResult };
