+++ packages/bridge-core/src/domains/auth/index.ts
@@ -1,0 +1,1 @@
+export type { SocialLoginResult } from './types';
