+++ packages/bridge-core/src/domains/auth/types.ts
@@ -1,0 +1,8 @@
+type SocialLoginResult = {
+  idToken: string;
+  email: string;
+  name?: string;
+  picture?: string;
+};
+
+export type { SocialLoginResult };
