+++ packages/bridge-core/src/domains/location/index.ts
@@ -1,0 +1,3 @@
+export type { LocationRPCSchema } from './rpc-schema';
+export type { LOCATION_ERROR_CODES } from './errors';
+export type { Accuracy, Location, PermissionStatus } from './types';
