+++ packages/bridge-core/src/domains/media/index.ts
@@ -1,0 +1,2 @@
+export type { PickImageParams, ImageAsset, PickImageResult } from './types';
+export type { MediaEventMap } from './events';
