+++ packages/bridge-core/src/domains/navigation/events.ts
@@ -1,0 +1,6 @@
+interface NavigationEventMap {
+  'nav.result': { txId: string; result: unknown };
+  'nav.cancel': { txId: string; reason?: string };
+}
+
+export type { NavigationEventMap };
