+++ packages/bridge-core/src/domains/navigation/index.ts
@@ -1,0 +1,2 @@
+export type { NavigationRPCSchema } from './rpc-schema';
+export type { NavigationEventMap } from './events';
