+++ packages/bridge-core/src/domains/safe-area/index.ts
@@ -1,0 +1,2 @@
+export type { SafeAreaRPCSchema } from './rpc-schema';
+export type { SafeAreaInsets } from './types';
