+++ packages/bridge-core/src/domains/safe-area/rpc-schema.ts
@@ -1,0 +1,11 @@
+import { METHODS } from '@/rpc';
+import type { SafeAreaInsets } from './types';
+
+interface SafeAreaRPCSchema {
+  [METHODS.getSafeAreaInsets]: {
+    params: {};
+    result: SafeAreaInsets;
+  };
+}
+
+export type { SafeAreaRPCSchema };
