+++ packages/bridge-core/src/domains/safe-area/types.ts
@@ -1,0 +1,9 @@
+type SafeAreaInsets = {
+  top: number;
+  bottom: number;
+  left: number;
+  right: number;
+};
+
+
+export type { SafeAreaInsets };
