+++ packages/bridge-core/src/domains/system/events.ts
@@ -1,0 +1,7 @@
+import type { ShareParams } from './types';
+
+export interface SystemEventMap {
+  'system.openExternalLink': { url: string };
+  'system.openSystemSetting': undefined;
+  'system.share': ShareParams;
+}
