+++ packages/bridge-core/src/domains/system/index.ts
@@ -1,0 +1,3 @@
+export type { SystemRPCSchema } from './rpc-schema';
+export type { CallPhoneResult, ShareParams } from './types';
+export type { SystemEventMap } from './events';
