+++ packages/bridge-core/src/domains/toast/events.ts
@@ -1,0 +1,12 @@
+interface ToastEventMap {
+  'ui.toast': {
+    id?: string;
+    title?: string;
+    description?: string;
+    duration?: number;
+    variant?: 'rounded' | 'square';
+    position?: 'top' | 'bottom' | 'bottom-above-nav';
+  };
+}
+
+export type { ToastEventMap };
