+++ packages/bridge-core/src/domains/toast/index.ts
@@ -1,0 +1,3 @@
+export type { ToastRPCSchema } from './rpc-schema';
+export type { ToastShape, ToastPosition, ToastType, ToastShowParams, ToastDismissParams, ToastClearParams } from './types';
+export type { ToastEventMap } from './events';
