+++ packages/bridge-core/src/domains/toast/rpc-schema.ts
@@ -1,0 +1,19 @@
+import { METHODS } from '@/rpc';
+import type { ToastShowParams, ToastDismissParams, ToastClearParams } from './types';
+
+interface ToastRPCSchema {
+  [METHODS.toastShow]: {
+    params: ToastShowParams;
+    result: void;
+  };
+  [METHODS.toastDismiss]: {
+    params: ToastDismissParams;
+    result: void;
+  };
+  [METHODS.toastClear]: {
+    params: ToastClearParams;
+    result: void;
+  };
+}
+
+export type { ToastRPCSchema };
