+++ packages/bridge-core/src/error/index.ts
@@ -1,0 +1,2 @@
+export type { BridgeErrorShape, BridgeErrorCode } from './types';
+export { BridgeException } from './BridgeError';
