+++ packages/bridge-core/src/events/index.ts
@@ -1,0 +1,1 @@
+export type { BridgeEventMap } from './map';
