+++ packages/bridge-core/src/index.ts
@@ -1,27 +1,25 @@
-export { METHODS, LOCATION_ERROR_CODES } from './methods';
-export type { BridgeEventMap } from './event-types';
-export { BRIDGE_VERSION, safeParse, makeId } from './utils';
-export type {
-  CallPhoneParams,
-  CallPhoneResult,
-  ShareParams,
-  ShareResult,
-  ToastShowParams,
-  ToastDismissParams,
-  ToastClearParams,
-  ToastShape,
-  ToastPosition,
-  ToastType,
-  SafeAreaInsets,
-  Accuracy,
-  Location,
-  PermissionStatus,
-  PickImageParams,
-  PickImageResult,
-  ImageAsset,
-  SocialLoginResult,
-  GetAppVersionResult,
-} from './methods';
-export type { RPCSchema, RPCMethod, ParamsOf, ResultOf } from './rpc-schema';
-export type { BridgeMessage, BridgeRequest, BridgeOk, BridgeError, BridgeEvent } from './types';
-export { BridgeException, makeBridgeError, type BridgeErrorCode, type BridgeErrorShape } from './error';
+export { METHODS } from '@/rpc';
+export { safeParse, makeId } from '@/utils';
+export { BRIDGE_VERSION } from '@/constants';
+export { type BridgeErrorShape, type BridgeErrorCode, BridgeException } from '@/error';
+export type { BridgeMessage, BridgeRequest } from '@/message';
+export type { RPCMethod, ParamsOf, ResultOf } from '@/rpc';
+export type { BridgeEventMap } from '@/events';
+
+export type { LOCATION_ERROR_CODES } from '@/domains/location';
+export type { CallPhoneResult } from '@/domains/system';
+export type { ShareParams } from '@/domains/system';
+export type { ToastShowParams, ToastDismissParams, ToastClearParams, ToastShape, ToastPosition, ToastType } from '@/domains/toast';
+export type { SafeAreaInsets } from '@/domains/safe-area';
+export type { Accuracy, Location, PermissionStatus } from '@/domains/location';
+export type { PickImageParams, ImageAsset, PickImageResult } from '@/domains/media';
+export type { GetAppVersionResult } from '@/domains/app-version';
+export type { SocialLoginResult } from '@/domains/auth';
+
+
+export type { Test };
+
+interface Test {
+  key: string;
+  preSignedUrl: string;
+}
