+++ packages/bridge-core/src/message/index.ts
@@ -1,0 +1,1 @@
+export type { BridgeMessage, BridgeRequest } from './protocol';
