--- packages/bridge-core/src/methods.ts
@@ -1,220 +1,0 @@
-const METHODS = {
-  getLatLng: 'device.getLatLng',
-  getSafeAreaInsets: 'device.getSafeAreaInsets',
-  getCurrentLocation: 'device.getCurrentLocation',
-  getLocationPermission: 'device.getLocationPermission',
-  requestLocationPermission: 'device.requestLocationPermission',
-  /** 기기 위치 서비스(GPS) ON/OFF 상태 확인 */
-  isLocationServiceEnabled: 'device.isLocationServiceEnabled',
-  /** 캐시된 마지막 위치 즉시 반환 (GPS 호출 없음) */
-  getLastKnownLocation: 'device.getLastKnownLocation',
-  callPhone: 'system.callPhone',
-  copyToClipboard: 'system.copyToClipboard',
-  share: 'system.share',
-  navPush: 'system.navPush',
-  navBack: 'system.navBack',
-  navReset: 'system.navReset',
-  navReplace: 'system.navReplace',
-  navSwitchTab: 'system.navSwitchTab',
-  naverOpenRoute: 'naver.openRoute',
-  toastShow: 'toast.show',
-  toastDismiss: 'toast.dismiss',
-  toastClear: 'toast.clear',
-  openExternalLink: 'system.openExternalLink',
-  openSettings: 'system.openSettings',
-  getAppVersion: 'system.getAppVersion',
-  kakaoLogin: 'auth.kakaoLogin',
-  googleLogin: 'auth.googleLogin',
-  appleLogin: 'auth.appleLogin',
-} as const;
-
-export type MethodName = (typeof METHODS)[keyof typeof METHODS];
-
-type CallPhoneParams = {
-  phoneNumber: string;
-};
-
-type CallPhoneResult = {
-  opened: boolean;
-  simulated?: boolean; // 시뮬레이터에서 실행된 경우
-};
-
-type ShareMandatory =
-  | {
-      url: string;
-      message?: string;
-    }
-  | {
-      url?: string;
-      message: string;
-    };
-
-type ShareParams = ShareMandatory & {
-  // iOS 전용 옵션
-  subject?: string; // 이메일 공유 시 제목
-  excludedActivityTypes?: string[]; // 제외할 액티비티 타입들
-  tintColor?: string; // 액티비티 색상
-
-  // Android 전용 옵션
-  title?: string; // 메시지 제목 (선택사항)
-  dialogTitle?: string; // 공유 다이얼로그 제목
-};
-
-type ShareResult = {
-  shared: boolean; // 공유 되었는지 여부
-  activityType?: string; // IOS에서 어떤 activity가 선택되었는지
-};
-
-type ToastShape = 'rounded' | 'square';
-type ToastPosition = 'top' | 'bottom' | 'bottom-above-nav';
-type ToastType = 'default' | 'success';
-
-type ToastShowParams = {
-  id?: string;
-  title?: string;
-  description?: string;
-  duration?: number; // ms
-  position?: ToastPosition;
-  shape?: ToastShape;
-  type?: ToastType;
-};
-type ToastDismissParams = { id?: string };
-type ToastClearParams = {};
-
-type SafeAreaInsets = {
-  top: number;
-  bottom: number;
-  left: number;
-  right: number;
-};
-
-type Accuracy = 'balanced' | 'high' | 'low';
-
-type Location = {
-  /**
-   * 자세한 위치 정보
-   */
-  coords: LocationCoords;
-  /**
-   * 위치가 업데이트된 시점의 타임스탬프
-   */
-  timestamp: number;
-};
-
-interface LocationCoords {
-  /**
-   * 위도
-   */
-  latitude: number;
-  /**
-   * 경도
-   */
-  longitude: number;
-  /**
-   * 정확도
-   */
-  accuracy: number | null;
-  /**
-   * 고도
-   */
-  altitude: number | null;
-  /**
-   * 고도 정확도
-   */
-  altitudeAccuracy: number | null;
-  /**
-   * 방향
-   */
-  heading: number | null;
-  /**
-   * 속도
-   */
-  speed: number | null;
-}
-
-type PermissionStatus = 'allowed' | 'denied' | 'undetermined';
-
-/**
- * 위치 에러 코드
- */
-const LOCATION_ERROR_CODES = {
-  /** 위치 권한 거부 */
-  PERMISSION_DENIED: 'LOCATION_PERMISSION_DENIED',
-  /** 위치 서비스(GPS) OFF */
-  SERVICE_DISABLED: 'LOCATION_SERVICE_DISABLED',
-  /** 위치 획득 불가 (실내, 지하 등) */
-  UNAVAILABLE: 'LOCATION_UNAVAILABLE',
-  /** 위치 획득 타임아웃 */
-  TIMEOUT: 'LOCATION_TIMEOUT',
-} as const;
-
-/** isLocationServiceEnabled 결과 */
-type IsLocationServiceEnabledResult = {
-  enabled: boolean;
-};
-
-/** getLastKnownLocation 파라미터 */
-type GetLastKnownLocationParams = {
-  maxAge?: number;
-};
-
-/** getLastKnownLocation 결과 */
-type GetLastKnownLocationResult = Location | null;
-
-type PickImageParams = {
-  source?: 'camera' | 'library';
-  allowsEditing?: boolean; // 이미지 편집 허용 여부
-  quality?: number; // 0~1 사이의 압축 품질
-  aspect?: [number, number]; // 종횡비 [가로, 세로]
-  mediaTypes?: 'images' | 'videos' | 'all'; // 미디어 타입
-  allowsMultipleSelection?: boolean; // 다중 선택 여부 (기본: false)
-  orderedSelection?: boolean; // 선택 순서 유지 여부 (iOS만 지원, 다중 선택 시)
-  selectionLimit?: number; // 선택 가능한 최대 개수 (0 = 무제한, 다중 선택 시)
-};
-
-type ImageAsset = {
-  preSignedUrl: string; // 업로드된 이미지 pre-signed URL
-  key: string; // 업로드된 이미지 키
-};
-
-type PickImageResult = {
-  cancelled: boolean; // 취소 여부
-  assets?: ImageAsset[]; // 업로드된 이미지 정보
-};
-
-type SocialLoginResult = {
-  idToken: string;
-  email: string;
-  name?: string;
-  picture?: string;
-};
-
-type GetAppVersionResult = {
-  version: string;
-};
-
-export { METHODS, LOCATION_ERROR_CODES };
-export type {
-  CallPhoneParams,
-  CallPhoneResult,
-  ShareParams,
-  ShareResult,
-  ToastShowParams,
-  ToastDismissParams,
-  ToastClearParams,
-  ToastShape,
-  ToastPosition,
-  ToastType,
-  SafeAreaInsets,
-  Accuracy,
-  Location,
-  PermissionStatus,
-  IsLocationServiceEnabledResult,
-  GetLastKnownLocationParams,
-  GetLastKnownLocationResult,
-  PickImageParams,
-  PickImageResult,
-  ImageAsset,
-  SocialLoginResult,
-  GetAppVersionResult,
-};
