--- packages/bridge-core/src/rpc-schema.ts
@@ -1,138 +1,0 @@
-import {
-  METHODS,
-  type CallPhoneParams,
-  type CallPhoneResult,
-  type ShareParams,
-  type ShareResult,
-  type ToastShowParams,
-  type ToastDismissParams,
-  type ToastClearParams,
-  type SafeAreaInsets,
-  type Accuracy,
-  type Location,
-  type PermissionStatus,
-  type GetAppVersionResult,
-  type IsLocationServiceEnabledResult,
-  type GetLastKnownLocationParams,
-  type GetLastKnownLocationResult,
-} from './methods';
-
-interface RPCSchema {
-  [METHODS.getLatLng]: {
-    params: {
-      accuracy?: 'balanced' | 'high' | undefined;
-    };
-    result: {
-      lat: number;
-      lng: number;
-    };
-  };
-  [METHODS.getSafeAreaInsets]: {
-    params: {};
-    result: SafeAreaInsets;
-  };
-  [METHODS.getCurrentLocation]: {
-    params: { accuracy?: Accuracy };
-    result: Location;
-  };
-  [METHODS.getLocationPermission]: {
-    params: {};
-    result: { status: PermissionStatus };
-  };
-  [METHODS.requestLocationPermission]: {
-    params: {};
-    result: { status: PermissionStatus; canAskAgain: boolean };
-  };
-  [METHODS.isLocationServiceEnabled]: {
-    params: {};
-    result: IsLocationServiceEnabledResult;
-  };
-  [METHODS.getLastKnownLocation]: {
-    params: GetLastKnownLocationParams;
-    result: GetLastKnownLocationResult;
-  };
-  [METHODS.openSettings]: {
-    params: {};
-    result: { opened: boolean };
-  };
-  [METHODS.callPhone]: {
-    params: CallPhoneParams;
-    result: CallPhoneResult;
-  };
-  [METHODS.copyToClipboard]: {
-    params: {
-      text: string;
-    };
-    result: {
-      copied: boolean;
-    };
-  };
-  [METHODS.share]: {
-    params: ShareParams;
-    result: ShareResult;
-  };
-  [METHODS.navPush]: {
-    params: {
-      name: string;
-      params: Record<string, unknown>;
-    };
-    result: {
-      pushed: boolean;
-    };
-  };
-  [METHODS.navBack]: {
-    params: {};
-    result: {
-      wentBack: boolean;
-    };
-  };
-  [METHODS.navReplace]: {
-    params: {
-      name: string;
-      params: Record<string, unknown>;
-    };
-    result: {
-      replaced: boolean;
-    };
-  };
-  [METHODS.navReset]: {
-    params: {
-      name: string;
-      params: Record<string, unknown>;
-    };
-    result: {
-      reset: boolean;
-    };
-  };
-  [METHODS.navSwitchTab]: {
-    params: {
-      pathname: string;
-      query?: Record<string, unknown>;
-    };
-    result: {
-      switched: boolean;
-    };
-  };
-  [METHODS.toastShow]: {
-    params: ToastShowParams;
-    result: void;
-  };
-  [METHODS.toastDismiss]: {
-    params: ToastDismissParams;
-    result: void;
-  };
-  [METHODS.toastClear]: {
-    params: ToastClearParams;
-    result: void;
-  };
-  [METHODS.getAppVersion]: {
-    params: {};
-    result: GetAppVersionResult;
-  };
-}
-
-type RPCMethod = keyof RPCSchema;
-type ParamsOf<K extends RPCMethod> = RPCSchema[K]['params'];
-type ResultOf<K extends RPCMethod> = RPCSchema[K]['result'];
-
-export type { RPCSchema, RPCMethod, ParamsOf, ResultOf };
