+++ packages/bridge-core/src/rpc/index.ts
@@ -1,0 +1,2 @@
+export { METHODS } from './methods';
+export type { RPCMethod, ParamsOf, ResultOf } from './schema';
