+++ packages/bridge-core/src/utils/id.ts
@@ -1,0 +1,5 @@
+function makeId() {
+  return crypto.randomUUID();
+}
+
+export { makeId };
