+++ packages/bridge-core/src/utils/json.ts
@@ -1,0 +1,9 @@
+function safeParse<T = unknown>(json: string): T | null {
+  try {
+    return JSON.parse(json) as T;
+  } catch {
+    return null;
+  }
+}
+
+export { safeParse };
