+++ packages/bridge-core/tsconfig.json
@@ -1,8 +1,18 @@
 {
   "extends": "@knockdog/typescript-config/base.json",
   "compilerOptions": {
-    "outDir": "./dist"
+    "outDir": "./dist",
+    "baseUrl": ".",
+    "paths": {
+      "@/rpc": ["./src/rpc"],
+      "@/domains/*": ["./src/domains/*"],
+      "@/error": ["./src/error"],
+      "@/message": ["./src/message"],
+      "@/utils": ["./src/utils"],
+      "@/constants": ["./src/constants"],
+      "@/events": ["./src/events"]
+    }
   },
-  "include": ["src/**/*", "../bridge-web/src/global.d.ts"],
+  "include": ["src/**/*"],
   "exclude": ["node_modules", "dist"]
 }
